backend/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
python-dotenv==1.2.1
requests==2.32.5

google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1